requirements.txt
Django==1.8.5
django-filter==0.13.0
djangorestframework==3.3.0
selenium==2.53.6
psycopg2==2.6.2
invoke==0.13.0
